requirements.txt
# AFA现代化依赖 - Python 3.10+

# CLI界面
click>=8.1.0

# 异步支持
asyncio-pool>=0.6.0

# 配置管理
PyYAML>=6.0
pydantic>=2.0.0

# 图像处理 (现代版本)
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
numpy>=1.24.0

# 机器学习和模型 (现代版本)
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
timm>=0.9.0

# 向量搜索和知识库
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # 如果需要GPU支持
torchvision>=0.15.0
scikit-learn>=1.3.0

# 向量检索和数据库
faiss-cpu>=1.7.4  # 使用CPU版本以提高兼容性
sqlalchemy>=2.0.0

# 数据科学工具
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM和向量相关
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# 开发和测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# 类型注解
typing-extensions>=4.5.0

# 日志
loguru>=0.7.0

# HTTP客户端
httpx>=0.24.0

# 进度条
tqdm>=4.65.0

# 数据验证
marshmallow>=3.20.0
transformers>=4.20.0

# 元数据和特征提取
exifread>=3.0.0
pywavelets>=1.4.0
scipy>=1.9.0

# 配置和工具
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 可选依赖（GPU加速）
# cupy-cuda11x>=11.0.0  # 根据CUDA版本选择
# tensorrt>=8.0.0       # TensorRT加速

# 开发和测试
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=5.0.0